Require rust >= 1.26 and drop libc_int128 conditional
[ resolve conflicts and add test skips that seem to be needed now -
  Trevor ]
