+++ src/unix/linux_like/linux/gnu/b64/aarch64/mod.rs
@@ -217,6 +217,12 @@ s! {
         __reserved: [[u64; 32]; 16],
     }
 
+    pub struct user_fpsimd_struct {
+        pub vregs: [::__uint128_t; 32],
+        pub fpsr: ::c_uint,
+        pub fpcr: ::c_uint,
+    }
+
     #[repr(align(8))]
     pub struct clone_args {
         pub flags: ::c_ulonglong,
@@ -962,15 +968,5 @@ cfg_if! {
     } else {
         mod lp64;
         pub use self::lp64::*;
-    }
-}
-
-cfg_if! {
-    if #[cfg(libc_int128)] {
-        mod int128;
-        pub use self::int128::*;
-    } else {
-        mod fallback;
-        pub use self::fallback::*;
     }
 }
